Support note responses
