+++ app/pages/notas/index.vue
@@ -2,7 +2,7 @@
 	<div id="notes" class="view --gap-none">
 		<section class="view-item --minHeightVh-100 --bgColor-light">
 			<div class="holder flx --flxColumn --flx-center --gap-50">
-				<div class="flx --flxColumn --flx-center --gap-30 --width-100">
+				<PaginationNotes :personal="personal">
 					<div class="txt --txtAlign-center">
 						<h1 class="--txtLineHeight-sm">Notas</h1>
 						<p v-if="SESSION.token" class="">Encuentra notas útiles o comparte una.</p>
@@ -22,122 +22,16 @@
 							>
 								Mis notas
 							</component>
-							<XamuActionButtonToggle
-								v-if="SESSION.canDevelop"
-								tooltip="Actualizar"
-								tooltip-position="right"
-								round
-								@click="emittedRefresh"
-							>
-								<XamuIconFa name="rotate-right" />
-								<XamuIconFa name="rotate-right" regular />
-							</XamuActionButtonToggle>
 						</div>
 					</div>
-					<XamuModal
-						id="new-note"
-						title="Nueva nota"
-						:save-button="{
-							title: !SESSION.token ? 'Ingresar a Cuna' : 'Publicar nota',
-						}"
-						class="--txtColor --txtAlign --minWidth-480:md --maxWidth-720:md"
-						invert-theme
-						@close="close"
-						@save="createNote"
-					>
-						<template #toggle="{ toggleModal }">
-							<div class="--width-100 --maxWidth-770">
-								<XamuBoxEditor
-									v-model="newNoteBody"
-									placeholder="Comparte algo que te guste..."
-									class="--width-100 --flx"
-								>
-									<template #submit>
-										<span
-											v-if="newNoteBody.length"
-											:class="{
-												'--txtColor-danger': remainingCharacters < 0,
-											}"
-											style="order: -1"
-										>
-											{{ remainingCharacters }}
-										</span>
-										<XamuSelect
-											v-if="SESSION.token"
-											v-model="newNotePublic"
-											:disabled="!newNoteBody"
-											:options="noteVisibilityOptions"
-										/>
-										<XamuActionButton
-											:disabled="!newNoteBody"
-											tooltip="Publicar nota"
-											tooltip-position="bottom"
-											round
-											@click="() => toggleModal(true)"
-										>
-											<span class="--hidden:md-inv">Publicar</span>
-											<XamuIconFa name="paper-plane" />
-										</XamuActionButton>
-									</template>
-								</XamuBoxEditor>
-							</div>
-						</template>
-						<template #default="{ model, invertedTheme }">
-							<div v-if="!SESSION.token" class="txt --gap-5">
-								<p>¡Espera un momento!</p>
-								<h3>Debes iniciar sesión para crear una nota</h3>
-							</div>
-							<XamuForm
-								v-else-if="model"
-								v-model="noteInputs"
-								v-model:invalid="invalidNote"
-								:theme="invertedTheme"
-							/>
-						</template>
-					</XamuModal>
-				</div>
-				<ClientOnly>
-					<template #fallback>Cargando notas...</template>
-					<XamuPaginationContent
-						v-slot="{ content }"
-						:page="notesPage"
-						url="api:instance:notes"
-						:defaults="{ page: true, personal: SESSION.token && personal }"
-						no-content-message="No hay notas disponibles, puedes crear una."
-						label="Cargando notas..."
-						class="flx --flxColumn --flx-start-center --maxWidth-770 --width-100 --gap-50"
-						hide-controls="single"
-						with-route
-						client
-						@refresh="emittedRefresh = $event"
-						@has-content="hasContent"
-					>
-						<ItemNote
-							v-for="(note, noteIndex) in content"
-							:key="note.id ?? noteIndex"
-							:note="note"
-							:hydrate-node="makeHydrateNode(noteIndex)"
-							:refresh="emittedRefresh"
-							class="--width-100"
-						/>
-					</XamuPaginationContent>
-				</ClientOnly>
+				</PaginationNotes>
 			</div>
 		</section>
 	</div>
 </template>
 
 <script setup lang="ts">
-	import { doc, type DocumentReference, getDoc } from "firebase/firestore";
-
-	import type { iGetPage, iInvalidInput, iPage, tFormInput } from "@open-xamu-co/ui-common-types";
-
-	import type { Note, NoteValues, NoteVoteRef } from "~/utils/types";
-
 	import { XamuActionButtonToggle, XamuActionLink } from "#components";
-	import type { NoteData } from "~~/functions/src/types/entities";
-
-	type HydrateNodes = (newContent: Note[] | null, newErrors?: unknown) => void;
 
 	/**
 	 * Note page
@@ -152,220 +46,10 @@
 		middleware: ["enabled"],
 	});
 
-	const { getResponse } = useFormInput();
-	const Swal = useSwal();
-	const INSTANCE = useInstanceStore();
 	const SESSION = useSessionStore();
-	const { $clientFirestore } = useNuxtApp();
-
-	// Form refs
-	const invalidNote = ref<iInvalidInput[]>([]);
-	const noteInputs = ref<tFormInput[]>(useNoteInputs());
-	const newNoteBody = ref<string>("");
-	const newNotePublic = ref<1 | 2 | 3>(1);
-
-	// Content refs
-	const emittedRefresh = ref<() => void>();
-	const emittedContent = ref<Note[] | null>();
-	const emittedHasContent = ref<boolean>();
-	const emittedHydrateNodes = ref<HydrateNodes>();
-	const deactivated = ref<boolean>(false);
 
 	/** Personal notes only */
 	const personal = ref<boolean>(false);
-
-	const remainingCharacters = computed(() => {
-		const limit = INSTANCE.config?.notesCharactersLimit ?? 4096;
-
-		return limit - newNoteBody.value.length;
-	});
-
-	const notesPage: iGetPage<Note> = (pagination) => {
-		return customFetch<iPage<Note> | undefined>("/api/instance/notes", {
-			query: pagination,
-			method: "POST",
-			credentials: "omit",
-			headers: { "Cache-Control": "no-store" },
-			cache: "no-store",
-		});
-	};
-
-	function close() {
-		invalidNote.value = [];
-		noteInputs.value = useNoteInputs();
-	}
-
-	async function createNote(closeModal: () => void, event: Event) {
-		if (!SESSION.token) {
-			closeModal();
-
-			return navigateTo({
-				path: "/ingresar",
-				query: { restricted: encodeURI("/notas"), rdr: "create-note" },
-			});
-		}
-
-		const { response, invalidInputs, withErrors, validationHadErrors, errors } =
-			await getResponse<{ id?: string }, NoteValues>(
-				async ({ name }) => {
-					try {
-						let publicValue: boolean | "UNLISTED" = true;
-
-						if (newNotePublic.value === 3) publicValue = "UNLISTED";
-						else if (newNotePublic.value === 2) publicValue = false;
-
-						// create note
-						const [data] = await useDocumentCreate<NoteData>(`${SESSION.path}/notes`, {
-							name,
-							body: newNoteBody.value.slice(
-								0,
-								INSTANCE.config?.notesCharactersLimit ?? 4096
-							),
-							public: publicValue,
-							keywords: name.trim().split(" "),
-						});
-
-						if (typeof data !== "object") return { errors: "Missing data" };
-
-						return { data };
-					} catch (errors) {
-						return { errors };
-					}
-				},
-				noteInputs.value,
-				event
-			);
-
-		invalidNote.value = invalidInputs;
-
-		const [createdNote, ...stream] = Array.isArray(response) ? response : [response];
-		let updatedNodes: Note[] | undefined;
-
-		if (!withErrors && createdNote) {
-			// Succesful request
-			Swal.fire({
-				title: "Nota creada exitosamente",
-				text: "Ya puedes encontrarla en las notas",
-				icon: "success",
-				willOpen() {
-					// Add new element at the beginning
-					if (typeof createdNote === "object" && createdNote.id) {
-						updatedNodes = emittedContent.value?.toSpliced(0, 0, createdNote);
-
-						// Hydration stream, do not await
-						Promise.allSettled(
-							stream.map(async (next) => {
-								const updated = await next;
-
-								// Bypass hydration
-								if (!updated || deactivated.value) return;
-
-								if (typeof updated === "object" && updated.id) {
-									const nodeIndex = updatedNodes?.findIndex(({ id }) => {
-										return id === updated.id;
-									});
-									const hydrateNode = makeHydrateNode(nodeIndex ?? -1);
-
-									// Hydrate if possible
-									hydrateNode({ ...createdNote, ...updated });
-								}
-							})
-						);
-					}
-
-					// Prefer hydration over refreshing
-					if (emittedHydrateNodes.value && updatedNodes) {
-						emittedHydrateNodes.value(updatedNodes);
-					} else if (!emittedHydrateNodes.value) emittedRefresh.value?.();
-
-					// Reset form
-					newNoteBody.value = "";
-					newNotePublic.value = 1;
-					closeModal?.();
-				},
-			});
-		} else if (!validationHadErrors) {
-			Swal.fire({
-				title: "¡Algo sucedió!",
-				text: "Ocurrió un error mientras creábamos la nota",
-				icon: "error",
-				target: event,
-			});
-
-			useAppLogger("pages:notes:createNote", errors);
-		}
-	}
-
-	/**
-	 * Handles content emission
-	 */
-	function hasContent(value: boolean, content?: Note[] | null, hydrateNodes?: HydrateNodes) {
-		emittedHasContent.value = value;
-		emittedContent.value = content;
-		emittedHydrateNodes.value = hydrateNodes;
-	}
-
-	function makeHydrateNode(nodeIndex: number) {
-		return (newNode: Note | null, _newErrors?: unknown) => {
-			if (!newNode) return;
-
-			// Replace the node with the updated one
-			const existingNode = emittedContent.value?.[nodeIndex];
-
-			if (nodeIndex > -1) {
-				const updatedNodes = emittedContent.value?.toSpliced(nodeIndex, 1, {
-					...existingNode,
-					...newNode,
-				});
-
-				// Hydrate node, fallback to refresh
-				if (updatedNodes && emittedHydrateNodes.value) {
-					emittedHydrateNodes.value(updatedNodes);
-				} else if (!emittedHydrateNodes.value) emittedRefresh.value?.();
-			}
-		};
-	}
-
-	// lifecycle
-	watch(
-		[emittedContent, () => SESSION.path],
-		async ([content = [], userPath], [oldContent = []]) => {
-			if (import.meta.server || !$clientFirestore) return;
-			// Bypass if same content
-			if (
-				!userPath ||
-				!content?.length ||
-				content[0]?.updatedAt === oldContent?.[0]?.updatedAt
-			) {
-				return;
-			}
-
-			// Map votes within notes
-			const hydratedNotes: Note[] = await Promise.all(
-				content.map(async (note) => {
-					// Get note vote
-					const id = `${note.id}/votes/${getDocumentId(userPath)}`;
-					const voteRef: DocumentReference<NoteVoteRef> = doc($clientFirestore, id);
-					const voteSnapshot = await getDoc(voteRef);
-					const { vote = 0 } = voteSnapshot?.data() || {};
-
-					return { ...note, vote };
-				})
-			);
-
-			// Hydrate nodes
-			emittedHydrateNodes.value?.(hydratedNotes);
-		},
-		{ immediate: true }
-	);
-	onActivated(() => {
-		if (deactivated.value) emittedRefresh.value?.();
-
-		deactivated.value = false;
-	});
-	onDeactivated(() => {
-		deactivated.value = true;
-	});
 </script>
 
 <style lang="scss">
@@ -376,14 +60,6 @@
 		*/
 		#notes.view li:has(> select#first) {
 			display: none;
-		}
-	}
-</style>
-
-<style scoped lang="scss">
-	@media only screen {
-		.x-editor-span {
-			grid-column: 1/ -1;
 		}
 	}
 </style>
